migrate prompt to use sessionconfiguration
